# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 自动化运维平台 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: string;
  username: string;
  role: string;
}

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  online: "online",
  offline: "offline",
} as const;

export interface Agent {
  id: string;
  name: string;
  hostname: string;
  os: string;
  ip: string;
  status: AgentStatus;
  version?: string;
  lastSeen: string;
  createdAt: string;
}

export interface AgentStats {
  total: number;
  online: number;
  offline: number;
  commandsToday: number;
}

export interface ExecuteCommandBody {
  command: string;
  /** Timeout in seconds */
  timeout?: number;
}

export type CommandStatus = (typeof CommandStatus)[keyof typeof CommandStatus];

export const CommandStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
  timeout: "timeout",
} as const;

export interface Command {
  id: string;
  agentId: string;
  agentName: string;
  command: string;
  status: CommandStatus;
  output?: string;
  exitCode?: number;
  createdAt: string;
  completedAt?: string;
}

export interface AuditLog {
  id: string;
  action: string;
  target: string;
  detail?: string;
  operator: string;
  createdAt: string;
}