# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 自动化运维平台 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: agents
    description: Agent management
  - name: commands
    description: Remote command execution
  - name: auditLogs
    description: Audit logs

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"

  /agents/stats:
    get:
      operationId: getAgentStats
      tags: [agents]
      summary: Get agent statistics
      responses:
        "200":
          description: Agent statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStats"

  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a specific agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete/unregister an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /agents/{agentId}/execute:
    post:
      operationId: executeCommand
      tags: [commands]
      summary: Execute a command on an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCommandBody"
      responses:
        "200":
          description: Command dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        "404":
          description: Agent not found or offline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agents/{agentId}/commands:
    get:
      operationId: listAgentCommands
      tags: [commands]
      summary: List command history for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Command history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"

  /commands:
    get:
      operationId: listAllCommands
      tags: [commands]
      summary: List all commands across all agents
      responses:
        "200":
          description: All commands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"

  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [auditLogs]
      summary: List audit logs
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - role

    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hostname:
          type: string
        os:
          type: string
        ip:
          type: string
        status:
          type: string
          enum: [online, offline]
        version:
          type: string
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - hostname
        - os
        - ip
        - status
        - lastSeen
        - createdAt

    AgentStats:
      type: object
      properties:
        total:
          type: integer
        online:
          type: integer
        offline:
          type: integer
        commandsToday:
          type: integer
      required:
        - total
        - online
        - offline
        - commandsToday

    ExecuteCommandBody:
      type: object
      properties:
        command:
          type: string
        timeout:
          type: integer
          description: Timeout in seconds
      required:
        - command

    Command:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        command:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, timeout]
        output:
          type: string
        exitCode:
          type: integer
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - agentName
        - command
        - status
        - createdAt

    AuditLog:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        target:
          type: string
        detail:
          type: string
        operator:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - target
        - operator
        - createdAt