# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 自动化运维平台 API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  role: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  role: zod.string(),
});

/**
 * @summary List all agents
 */
export const ListAgentsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  hostname: zod.string(),
  os: zod.string(),
  ip: zod.string(),
  status: zod.enum(["online", "offline"]),
  version: zod.string().optional(),
  lastSeen: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const ListAgentsResponse = zod.array(ListAgentsResponseItem);

/**
 * @summary Get agent statistics
 */
export const GetAgentStatsResponse = zod.object({
  total: zod.number(),
  online: zod.number(),
  offline: zod.number(),
  commandsToday: zod.number(),
});

/**
 * @summary Get a specific agent
 */
export const GetAgentParams = zod.object({
  agentId: zod.coerce.string(),
});

export const GetAgentResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  hostname: zod.string(),
  os: zod.string(),
  ip: zod.string(),
  status: zod.enum(["online", "offline"]),
  version: zod.string().optional(),
  lastSeen: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete/unregister an agent
 */
export const DeleteAgentParams = zod.object({
  agentId: zod.coerce.string(),
});

export const DeleteAgentResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Execute a command on an agent
 */
export const ExecuteCommandParams = zod.object({
  agentId: zod.coerce.string(),
});

export const ExecuteCommandBody = zod.object({
  command: zod.string(),
  timeout: zod.number().optional().describe("Timeout in seconds"),
});

export const ExecuteCommandResponse = zod.object({
  id: zod.string(),
  agentId: zod.string(),
  agentName: zod.string(),
  command: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed", "timeout"]),
  output: zod.string().optional(),
  exitCode: zod.number().optional(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().optional(),
});

/**
 * @summary List command history for an agent
 */
export const ListAgentCommandsParams = zod.object({
  agentId: zod.coerce.string(),
});

export const ListAgentCommandsResponseItem = zod.object({
  id: zod.string(),
  agentId: zod.string(),
  agentName: zod.string(),
  command: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed", "timeout"]),
  output: zod.string().optional(),
  exitCode: zod.number().optional(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().optional(),
});
export const ListAgentCommandsResponse = zod.array(
  ListAgentCommandsResponseItem,
);

/**
 * @summary List all commands across all agents
 */
export const ListAllCommandsResponseItem = zod.object({
  id: zod.string(),
  agentId: zod.string(),
  agentName: zod.string(),
  command: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed", "timeout"]),
  output: zod.string().optional(),
  exitCode: zod.number().optional(),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().optional(),
});
export const ListAllCommandsResponse = zod.array(ListAllCommandsResponseItem);

/**
 * @summary List audit logs
 */
export const ListAuditLogsResponseItem = zod.object({
  id: zod.string(),
  action: zod.string(),
  target: zod.string(),
  detail: zod.string().optional(),
  operator: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAuditLogsResponse = zod.array(ListAuditLogsResponseItem);